+++ examples/random_rounding/main.go
@@ -2,6 +2,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"math"
 	"math/rand"
@@ -69,6 +70,9 @@ func (h *randomRoundingHeur) Execute(model scip.Model, heur scip.HeuristicPlugin
 
 	// Try to add the rounded solution
 	if err := model.AddSol(&sol); err != nil {
+		if !errors.Is(err, scip.SolErrorInfeasible) {
+			panic(err) // a SCIP failure or a constraint handler panic, not a rejected guess
+		}
 		fmt.Println("-- RandomRoundingHeur: Failed to add solution to the model.")
 		return scip.HeurResultNoSolFound
 	}